Remove webpack.common.js
