+++ Content/default/SAFE.App.sln
@@ -10,7 +10,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		package-lock.json = package-lock.json
 		package.json = package.json
 		webpack.config.js = webpack.config.js
-		webpack.common.js = webpack.common.js
 		README.md = README.md
 	EndProjectSection
 EndProject
